Animations & Content
Made all of the slides, put in about me and work history content.
Got the fade in and out animations to work
removed the scroll bar for immersion

Need to add in the graphics and some content into sceneThree.

Then the final cleanup of the code, implementing better modular code using better functions etc.